Steal GNB's homework for a generic core disengage GCD module
